cleanup

+++ include/maximilian/chorus.h
@@ -2,8 +2,8 @@
 
 #include "core.h"
 #include "delayline.h"
-#include "oscillator.h"
 #include "filter.h"
+#include "oscillator.h"
 
 namespace maximilian {
 
@@ -15,8 +15,8 @@ public:
   // feedback = 0 - 1
   // speed = lfo speed in Hz, 0.0001 - 10 sounds good
   // depth = 0 - 1
-  double process(double input, unsigned int delay,
-                double feedback, double speed, double depth);
+  float process(float input, unsigned int delay, float feedback, float speed,
+                float depth);
 
 private:
   Delayline dl, dl2;
