+++ include/maximilian/core.h
@@ -2,7 +2,7 @@
 
 namespace maximilian {
 
-const double TWOPI = 6.283185307179586476925286766559;
+const float TWOPI = 6.283185307179586476925286766559;
 
 struct Context final {
   int sampleRate = 44100;
@@ -15,6 +15,5 @@ struct Processor {
   Context &context;
 };
 
-double mtof(int midinote);
-
+float mtof(int midinote);
 }
