+++ include/maximilian/delayline.h
@@ -8,16 +8,16 @@ class Delayline final : public Processor {
 
 public:
   Delayline(Context &context);
-  double process(double input, int size, double feedback);
-  double process(double input, int size, double feedback, int position);
+  float process(float input, int size, float feedback);
+  float process(float input, int size, float feedback, int position);
 
 private:
-  double frequency;
+  float frequency;
   int phase;
-  double startphase;
-  double endphase;
-  double output;
-  double memory[88200];
+  float startphase;
+  float endphase;
+  float output;
+  float memory[88200];
 };
 
 } // namespace maximilian
