+++ include/maximilian/dynamic.h
@@ -9,21 +9,21 @@ class Dynamic : public Processor {
 public:
   Dynamic(Context &context);
 
-  double gate(double input, double threshold = 0.9, long holdtime = 1,
-              double attack = 1, double release = 0.9995);
+  float gate(float input, float threshold = 0.9f, long holdtime = 1,
+             float attack = 1, float release = 0.9995f);
 
-  double compressor(double input, double ratio, double threshold = 0.9,
-                    double attack = 1, double release = 0.9995);
+  float compressor(float input, float ratio, float threshold = 0.9f,
+                   float attack = 1, float release = 0.9995f);
 
 private:
-  double input;
-  double ratio;
-  double currentRatio;
-  double threshold;
-  double output;
-  double attack;
-  double release;
-  double amplitude;
+  float input;
+  float ratio;
+  float currentRatio;
+  float threshold;
+  float output;
+  float attack;
+  float release;
+  float amplitude;
   long holdtime;
   long holdcount;
   int attackphase, holdphase, releasephase;
@@ -32,8 +32,8 @@ private:
 struct Compressor final : public Dynamic {
   Compressor(Context &context) : Dynamic(context) {}
 
-  double process(double input, double ratio, double threshold = 0.9,
-                 double attack = 1, double release = 0.9995) {
+  float process(float input, float ratio, float threshold = 0.9f,
+                float attack = 1, float release = 0.9995f) {
     return compressor(input, ratio, threshold, attack, release);
   }
 };
@@ -41,8 +41,8 @@ struct Compressor final : public Dynamic {
 struct Gate final : public Dynamic {
   Gate(Context &context) : Dynamic(context) {}
 
-  double process(double input, double threshold = 0.9, long holdtime = 1,
-                 double attack = 1, double release = 0.9995) {
+  float process(float input, float threshold = 0.9f, long holdtime = 1,
+                float attack = 1, float release = 0.9995f) {
     return gate(input, threshold, holdtime, attack, release);
   }
 };
