+++ include/maximilian/envelope.h
@@ -8,18 +8,18 @@ class Envelope final : public Processor {
 
 public:
   Envelope(Context &context);
-  double line(int numberofsegments, double segments[100]);
-  void trigger(int index, double amp);
+  float line(int numberofsegments, float segments[100]);
+  void trigger(int index, float amp);
 
 private:
-  double period;
-  double output;
-  double startval;
-  double currentval;
-  double nextval;
+  float period;
+  float output;
+  float startval;
+  float currentval;
+  float nextval;
   int isPlaying;
   int valindex;
-  double amplitude;
+  float amplitude;
 };
 
 } // namespace maximilian
