+++ include/maximilian/envelopefollower.h
@@ -10,13 +10,13 @@ namespace maximilian {
 class EnvelopeFollower final : public Processor {
 public:
   EnvelopeFollower(Context &context);
-  void setAttack(double attackMS);
-  void setRelease(double releaseMS);
-  double process(double input);
+  void setAttack(float attackMS);
+  void setRelease(float releaseMS);
+  float process(float input);
   void reset();
 
 private:
-  double attack, release, env;
+  float attack, release, env;
 };
 
 } // namespace maximilian
