+++ include/maximilian/flanger.h
@@ -15,8 +15,8 @@ public:
   // feedback = 0 - 1
   // speed = lfo speed in Hz, 0.0001 - 10 sounds good
   // depth = 0 - 1
-  double process(double input, unsigned int delay,
-                double feedback, double speed, double depth);
+  float process(float input, unsigned int delay, float feedback, float speed,
+                float depth);
 
 private:
   Delayline dl;
