+++ src/distortion.cpp
@@ -3,13 +3,13 @@
 
 namespace maximilian {
 
-double fastatan(double x) { return (x / (1.0 + 0.28 * (x * x))); }
+float fastatan(float x) { return (x / (1.0 + 0.28 * (x * x))); }
 
-double atanDist(double in, double shape) {
+float atanDist(float in, float shape) {
   return (1.0 / std::atan(shape)) * std::atan(in * shape);
 }
 
-double fastAtanDist(double in, double shape) {
+float fastAtanDist(float in, float shape) {
   return (1.0 / fastatan(shape)) * fastatan(in * shape);
 }
