+++ src/envelopefollower.cpp
@@ -9,17 +9,17 @@ EnvelopeFollower::EnvelopeFollower(Context &context) : Processor(context) {
   env = 0;
 }
 
-void EnvelopeFollower::setAttack(double attackMS) {
+void EnvelopeFollower::setAttack(float attackMS) {
   attack =
       std::pow(0.01, 1.0 / (attackMS * context.sampleRate * 0.001));
 }
 
-void EnvelopeFollower::setRelease(double releaseMS) {
+void EnvelopeFollower::setRelease(float releaseMS) {
   release =
       std::pow(0.01, 1.0 / (releaseMS * context.sampleRate * 0.001));
 }
 
-double EnvelopeFollower::process(double input) {
+float EnvelopeFollower::process(float input) {
   input = fabs(input);
   
   if (input > env)
