+++ src/flanger.cpp
@@ -10,9 +10,9 @@ Flanger::Flanger(Context &context)
 // feedback = 0 - 1
 // speed = lfo speed in Hz, 0.0001 - 10 sounds good
 // depth = 0 - 1
-double Flanger::process(double input, unsigned int delay,
-                       double feedback, double speed,
-                       double depth) {
+float Flanger::process(float input, unsigned int delay,
+                       float feedback, float speed,
+                       float depth) {
   auto lfoVal = lfo.triangle(speed);
   auto output = dl.process(input, delay + (lfoVal * depth * delay) + 1, feedback);
   auto normalise = (1 - std::fabs(output));
